+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added `Debug` for all FTDI device structures.
 
+### Changed
+- Changed the `with_description` argument name from `serial_number` to
+  `description`.
+
 ## [0.25.1] - 2021-03-13
 ### Fixed
 - Fixed `list_devices_fs` returning an `Err` when FTDI devices with invalid